+++ packages/avalanche-module/src/handlers/avalanche-send-transaction/avalanche-send-transaction.test.ts
@@ -406,7 +406,7 @@ describe('avalanche_sendTransaction handler', () => {
       const params = testParams(testRequestParams);
       const response = await avalancheSendTransaction(params);
 
-      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true, isDevnet: false });
+      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true });
 
       expect(issueTxHexMock).toHaveBeenCalledWith(testSignedTxHash, 'AVM');
 
@@ -423,7 +423,7 @@ describe('avalanche_sendTransaction handler', () => {
       });
       const response = await avalancheSendTransaction(params);
 
-      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true, isDevnet: false });
+      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true });
 
       expect(issueTxHexMock).toHaveBeenCalledWith(testSignedTxHash, 'AVM');
 
@@ -437,7 +437,7 @@ describe('avalanche_sendTransaction handler', () => {
       const params = testParams({ transactionHex: '0x000142', chainAlias: 'C' });
       const response = await avalancheSendTransaction(params);
 
-      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true, isDevnet: false });
+      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true });
 
       expect(issueTxHexMock).toHaveBeenCalledWith(testSignedTxHash, 'EVM');
 
@@ -451,7 +451,7 @@ describe('avalanche_sendTransaction handler', () => {
       const params = testParams(testRequestParams);
       const response = await avalancheSendTransaction(params);
 
-      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true, isDevnet: false });
+      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true });
 
       expect(response).toStrictEqual({ result: testTxHash });
 
@@ -468,7 +468,7 @@ describe('avalanche_sendTransaction handler', () => {
 
       const response = await avalancheSendTransaction(params);
 
-      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true, isDevnet: false });
+      expect(mockGetProvider).toHaveBeenCalledWith({ isTestnet: true });
 
       expect(response).toStrictEqual({ result: testTxHash });