+++ packages/types/src/network-fee.ts
@@ -1,3 +1,16 @@
+import type { Hex } from './common';
+
+type SuggestedGasPriceOption = {
+  maxFeePerGas: Hex;
+  maxPriorityFeePerGas: Hex;
+};
+
+export type SuggestGasPriceOptionsResponse = {
+  slow: SuggestedGasPriceOption;
+  normal: SuggestedGasPriceOption;
+  fast: SuggestedGasPriceOption;
+};
+
 export type NetworkFees = {
   low: { maxFeePerGas: bigint; maxPriorityFeePerGas?: bigint };
   medium: { maxFeePerGas: bigint; maxPriorityFeePerGas?: bigint };